% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf1.R
\name{cf1.params.init}
\alias{cf1.params.init}
\title{Simple parameter estimation for PH}
\usage{
cf1.params.init(
  n,
  data,
  shape.init = c(1, 4, 16, 64, 256, 1024),
  scale.init = c(0.5, 1, 2),
  maxiter.init = 5,
  verbose = FALSE
)
}
\arguments{
\item{n}{An integer for the number of phases}

\item{data}{A data.frame for fault data. It should be an instance of Rsrat.faultdata}

\item{scale.init}{A numeric vector for scale}

\item{maxiter.init}{An integer of emsteps}

\item{verbose}{A logic to display the results of trials}

\item{diff.init}{A numeric vector for shape}
}
\value{
List of CF1 parameters.
}
\description{
Determine the parameters of CF1 from data.
Try all candidates with several EM steps and select the best one.
}
