% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf1.R
\name{cf1.params.linear}
\alias{cf1.params.linear}
\title{Simple parameter estimation for PH}
\usage{
cf1.params.linear(n, scale, shape)

cf1.params.linear(n, scale, shape)
}
\arguments{
\item{n}{An integer for the number of phases}

\item{scale}{A numeric vector of scale. This corresponds to the mean of CF1.}

\item{shape}{A numeric vector of shape. This is a parameter to indicate the
difference between the maximum/minimum rates.}
}
\value{
List of CF1 parameters.

List of CF1 parameters.
}
\description{
Determine the parameters of CF1 from data.
The rate increases linearly.

Determine the parameters of CF1 from data.
The rate increases linearly.
}
