% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cf1.R
\name{C_cf1pdf}
\alias{C_cf1pdf}
\alias{C_cf1cdf}
\alias{C_cf1sample}
\alias{C_cf1reform}
\alias{C_cf1sojourn}
\alias{cf1}
\alias{dcf1}
\alias{pcf1}
\alias{qcf1}
\alias{rcf1}
\title{Phase-Type distribution (Canonical)}
\usage{
C_cf1pdf(dx, alpha, rate, eps = 1e-08, ufactor = 1.01, log = FALSE)

C_cf1cdf(dx, alpha, rate, eps = 1e-08, ufactor = 1.01, lower = TRUE,
  log = FALSE)

C_cf1sample(n, alpha, rate)

C_cf1reform(alpha, rate)

C_cf1sojourn(alpha, rate, f, b, t, eps = 1e-08, ufactor = 1.01)

dcf1(x, alpha = 1, rate = 1, log = FALSE, eps = 1e-08,
  unif.factor = 1.01)

pcf1(q, alpha = 1, rate = 1, lower.tail = TRUE, log.p = FALSE,
  eps = 1e-08, unif.factor = 1.01)

qcf1(p, alpha = 1, rate = 1, lower.tail = TRUE, log.p = FALSE)

rcf1(n, alpha = 1, rate = 1, scramble = TRUE)
}
\arguments{
\item{dx}{A numeric value of time differences}

\item{alpha}{A vector of initial probabilities}

\item{rate}{A vector of transition rates}

\item{eps}{A value of tolerance error}

\item{ufactor}{A value of uniformization factor}

\item{log, log.p}{A logical; if TRUE, the probability p is given as log(p).}

\item{x, q}{A numeric vector of quantiles.}

\item{lower.tail}{A logical; if TRUE, the probability is P[X <= x], otherwise, P[X > x].}

\item{p}{A numeric vector of probabilities.}

\item{alpha}{A numeric vector for initial probabilities.}

\item{rate}{A numeric vector for transition rates to the next stage
(diagonal entries of infinitesimal generator matrix).}
}
\value{
'dcf1' gives the desity, 'pcf1' gives the distribution, and
'qcf1' gives the quantile function.
}
\description{
Density, distribution function, and quantile function for
canonical phase-type distribution.
}
\details{
The PH distribution has the density distribution function
F(x) = alpha exp(T x) tau where alpha is an initial probability vector,
T is an infinitesimal generator matrix and tau is an exit rate vector.

The canonical form 1 (CF1) is a special case of PH distribution, in which
The infinitesimal generator consists of only diagonal and superdiagonal entries, i.e.,
The phase moves to only the next stage and the last state is the absorbing state of
underlying continuous-time Markov chain. In CF1, the parameters are dominated only
by the initial probability vector and the transtion rates to the next stage. In addition,
transition rates of CF1 increases with the stage processes.
}
