% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf1.R
\name{cf1llf}
\alias{cf1llf}
\title{llf for CF1}
\usage{
cf1llf(data, omega, alpha, rate)
}
\arguments{
\item{data}{An object of faultdata}

\item{omega}{A parameter for the total number of faults}

\item{alpha}{A numeric vector for initial probabilities.}

\item{rate}{A numeric vector for transition rates.}
}
\value{
llf
}
\description{
Compute the llf of CF1
}
