% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf1.R
\name{cf1moment}
\alias{cf1moment}
\title{Moments of CF1}
\usage{
cf1moment(n, alpha, rate)
}
\arguments{
\item{n}{An integer of the degree of moment}

\item{alpha}{A numeric vector for initial probabilities.}

\item{rate}{A numeric vector for transition rates.}
}
\value{
A value of moment
}
\description{
Compute the moments of CF1
}
