% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphsrm.R
\name{cphsrm}
\alias{cphsrm}
\title{Software reliability model}
\usage{
cphsrm(phase)
}
\arguments{
\item{phase}{A numeric vector indicating the number of phases}
}
\value{
A class of CF1 (\code{\link{CPHSRM}}).
}
\description{
Generate a software reliability model.
}
\details{
Return a class of CPHSRM
}
