% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_cf1_emstep}
\alias{em_cf1_emstep}
\title{EM step}
\usage{
em_cf1_emstep(params, data, eps = 1e-08, ufactor = 1.01)
}
\arguments{
\item{params}{A numeric vector for model parameters}

\item{data}{A faultdata}

\item{eps}{A numeric value for tolerance error.}

\item{ufactor}{A numeric value indicating the uniformization factor}
}
\value{
A list with an updated parameter vector (param),
absolute difference of parameter vector (pdiff),
log-likelihood function for a given parameter vector (llf),
the number of total faults (total).
}
\description{
Execute one EM step for software reliability models.
}
