% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Rphsrm}
\alias{Rphsrm}
\alias{Rphsrm-package}
\title{Rphsrm: A Software Reliability Assessment Tool with PH distribution}
\description{
This package provides estimation programs for software reliability growth
models based on NHPP whose distribution is defined by phase-type distribution.
}
