% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_cf1srm.R
\docType{class}
\name{CPHSRM}
\alias{CPHSRM}
\title{Class for phase-type NHPP-based software reliability model (canonical form 1)}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for NHPP-based software reliability model.
}
\description{
Class for phase-type NHPP-based software reliability model (canonical form 1)

Class for phase-type NHPP-based software reliability model (canonical form 1)
}
\section{Super class}{
\code{\link[Rsrat:NHPP]{Rsrat::NHPP}} -> \code{CPHSRM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CPHSRM$new()}}
\item \href{#method-omega}{\code{CPHSRM$omega()}}
\item \href{#method-alpha}{\code{CPHSRM$alpha()}}
\item \href{#method-rate}{\code{CPHSRM$rate()}}
\item \href{#method-print}{\code{CPHSRM$print()}}
\item \href{#method-get_params}{\code{CPHSRM$get_params()}}
\item \href{#method-init_params}{\code{CPHSRM$init_params()}}
\item \href{#method-set_params}{\code{CPHSRM$set_params()}}
\item \href{#method-set_omega}{\code{CPHSRM$set_omega()}}
\item \href{#method-em}{\code{CPHSRM$em()}}
\item \href{#method-llf}{\code{CPHSRM$llf()}}
\item \href{#method-clone}{\code{CPHSRM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="cmtbf">}\href{../../Rsrat/html/NHPP.html#method-cmtbf}{\code{Rsrat::NHPP$cmtbf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="comp_error">}\href{../../Rsrat/html/NHPP.html#method-comp_error}{\code{Rsrat::NHPP$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="dmvf">}\href{../../Rsrat/html/NHPP.html#method-dmvf}{\code{Rsrat::NHPP$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="ffp">}\href{../../Rsrat/html/NHPP.html#method-ffp}{\code{Rsrat::NHPP$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="imtbf">}\href{../../Rsrat/html/NHPP.html#method-imtbf}{\code{Rsrat::NHPP$imtbf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="intensity">}\href{../../Rsrat/html/NHPP.html#method-intensity}{\code{Rsrat::NHPP$intensity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="inv_mvf">}\href{../../Rsrat/html/NHPP.html#method-inv_mvf}{\code{Rsrat::NHPP$inv_mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="median">}\href{../../Rsrat/html/NHPP.html#method-median}{\code{Rsrat::NHPP$median()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="mvf">}\href{../../Rsrat/html/NHPP.html#method-mvf}{\code{Rsrat::NHPP$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="reliab">}\href{../../Rsrat/html/NHPP.html#method-reliab}{\code{Rsrat::NHPP$reliab()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="residual">}\href{../../Rsrat/html/NHPP.html#method-residual}{\code{Rsrat::NHPP$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="sample">}\href{../../Rsrat/html/NHPP.html#method-sample}{\code{Rsrat::NHPP$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="set_data">}\href{../../Rsrat/html/NHPP.html#method-set_data}{\code{Rsrat::NHPP$set_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The number of total faults.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$new(n, omega = 1, alpha = rep(1/n, n), rate = rep(1, n))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{An integer for the number of phases.}

\item{\code{omega}}{A value for the number of total faults.}

\item{\code{alpha}}{A vector of the initial probabilities.}

\item{\code{rate}}{A vector of transition rates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega"></a>}}
\if{latex}{\out{\hypertarget{method-omega}{}}}
\subsection{Method \code{omega()}}{
The number of total faults.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$omega()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of total faults.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-alpha}{}}}
\subsection{Method \code{alpha()}}{
The initial probability vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$alpha()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The initial probability vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rate"></a>}}
\if{latex}{\out{\hypertarget{method-rate}{}}}
\subsection{Method \code{rate()}}{
The rate vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$rate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The rate vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$print(digits = max(3, getOption("digits") - 3), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{An integer to determine the number of digits for print.}

\item{\code{...}}{Others}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-get_params}{}}}
\subsection{Method \code{get_params()}}{
Make a flatten parameter vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$get_params(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{A vector of parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_params"></a>}}
\if{latex}{\out{\hypertarget{method-init_params}{}}}
\subsection{Method \code{init_params()}}{
Set initial parameters from a given data.
This is used to set the initial value for
the fitting algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$init_params(
  data,
  shape.init = c(1, 4, 16, 64, 256, 1024),
  scale.init = c(0.5, 1, 2),
  maxiter.init = 5,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.}

\item{\code{shape.init}}{A list of shapes.}

\item{\code{scale.init}}{A list of scale fraction.}

\item{\code{maxiter.init}}{An integer for the number of iterations.}

\item{\code{verbose}}{A Boolean whether the detailed information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
Set model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$set_params(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_omega"></a>}}
\if{latex}{\out{\hypertarget{method-set_omega}{}}}
\subsection{Method \code{set_omega()}}{
Set omega parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$set_omega(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A value of omega.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-em"></a>}}
\if{latex}{\out{\hypertarget{method-em}{}}}
\subsection{Method \code{em()}}{
Execute an EM step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$em(params, data, eps = 1e-08, ufactor = 1.01)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Parameters.}

\item{\code{data}}{Data.}

\item{\code{eps}}{A value for tolerance error.}

\item{\code{ufactor}}{A value for a factor to make uniformed kernel.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the following
\describe{
\item{param}{Updated parameters.}
\item{pdiff}{Absolute difference of parameter vector.}
\item{llf}{Log-likelihood function for a given parameter vector.}
\item{total}{The number of total faults.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-llf"></a>}}
\if{latex}{\out{\hypertarget{method-llf}{}}}
\subsection{Method \code{llf()}}{
The log-likelihood function for a given data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$llf(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Log-likelihood function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
