% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_cf1srm.R
\docType{class}
\name{CPHSRM}
\alias{CPHSRM}
\title{Class for phase-type NHPP-based software reliability model (canonical form 1)}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for NHPP-based software reliability model.
}
\description{
Class for phase-type NHPP-based software reliability model (canonical form 1)

Class for phase-type NHPP-based software reliability model (canonical form 1)
}
\section{Methods}{

\describe{
  \item{\code{print()}}{This method prints model parameters.}
  \item{\code{omega()}}{This method returns the number of total faults.}
  \item{\code{mvf(t)}}{This method returns the mean value function at time t.}
  \item{\code{dmvf(t)}}{This method returns the mean value function on discrete time domain.}
  \item{\code{inv_mvf(x)}}{This method returns the time at which the mean value function attains x.}
  \item{\code{intensity(t)}}{This method returns the intensity function at time t.}
  \item{\code{reliab(t, s)}}{This method returns the software reliability at time t from the orign s.}
  \item{\code{residual(t)}}{This method returns the expected residual number of faults at time t.}
  \item{\code{ffp(t)}}{This method returns the fault-free probability at time t.}
  \item{\code{imtbf(t)}}{This method returns the instantaneous MTBF at time t.}
  \item{\code{cmtbf(t)}}{This method returns the cumulative MTBF at time t.}
  \item{\code{median(s, p = 0.5)}}{This method returns the time at which the software reliability attains the proability p from the orign s.}
  \item{\code{get_params(params)}}{This method returns a flatten parameter vector.}
  \item{\code{init_params(data)}}{This method changes the model parameters based on a given data. This is used to set the initial value for the fitting algorithm.}
  \item{\code{set_params(params)}}{This method sets the model parameters.}
  \item{\code{em(params, data)}}{This method returns a list with an updated parameter vector (param),
         absolute difference of parameter vector (pdiff),
         log-likelihood function for a given parameter vector (llf),
         the number of total faults (total) via EM algorithm for a given data.
         \emph{eps} is the tolerance error for uniformization.
         \emph{ufactor} is a uniformization factor.}
  \item{\code{llf(data)}}{This method returns the log-likelihood function for a given data.}
}
}

\section{Super class}{
\code{\link[Rsrat:NHPP]{Rsrat::NHPP}} -> \code{CPHSRM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character string for the name of model.}

\item{\code{params}}{A list of the model parameters.}

\item{\code{df}}{An integer for the degrees of freedom of the model.}

\item{\code{data}}{Data to esimate parameters.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{A character string for the name of model.}

\item{\code{params}}{A list of the model parameters.}

\item{\code{df}}{An integer for the degrees of freedom of the model.}

\item{\code{data}}{Data to esimate parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CPHSRM$new()}}
\item \href{#method-omega}{\code{CPHSRM$omega()}}
\item \href{#method-alpha}{\code{CPHSRM$alpha()}}
\item \href{#method-rate}{\code{CPHSRM$rate()}}
\item \href{#method-print}{\code{CPHSRM$print()}}
\item \href{#method-get_params}{\code{CPHSRM$get_params()}}
\item \href{#method-init_params}{\code{CPHSRM$init_params()}}
\item \href{#method-set_params}{\code{CPHSRM$set_params()}}
\item \href{#method-em}{\code{CPHSRM$em()}}
\item \href{#method-llf}{\code{CPHSRM$llf()}}
\item \href{#method-clone}{\code{CPHSRM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="cmtbf">}\href{../../Rsrat/html/NHPP.html#method-cmtbf}{\code{Rsrat::NHPP$cmtbf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="comp_error">}\href{../../Rsrat/html/NHPP.html#method-comp_error}{\code{Rsrat::NHPP$comp_error()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="dmvf">}\href{../../Rsrat/html/NHPP.html#method-dmvf}{\code{Rsrat::NHPP$dmvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="ffp">}\href{../../Rsrat/html/NHPP.html#method-ffp}{\code{Rsrat::NHPP$ffp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="imtbf">}\href{../../Rsrat/html/NHPP.html#method-imtbf}{\code{Rsrat::NHPP$imtbf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="intensity">}\href{../../Rsrat/html/NHPP.html#method-intensity}{\code{Rsrat::NHPP$intensity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="inv_mvf">}\href{../../Rsrat/html/NHPP.html#method-inv_mvf}{\code{Rsrat::NHPP$inv_mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="median">}\href{../../Rsrat/html/NHPP.html#method-median}{\code{Rsrat::NHPP$median()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="mvf">}\href{../../Rsrat/html/NHPP.html#method-mvf}{\code{Rsrat::NHPP$mvf()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="reliab">}\href{../../Rsrat/html/NHPP.html#method-reliab}{\code{Rsrat::NHPP$reliab()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="residual">}\href{../../Rsrat/html/NHPP.html#method-residual}{\code{Rsrat::NHPP$residual()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="sample">}\href{../../Rsrat/html/NHPP.html#method-sample}{\code{Rsrat::NHPP$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="Rsrat" data-topic="NHPP" data-id="set_data">}\href{../../Rsrat/html/NHPP.html#method-set_data}{\code{Rsrat::NHPP$set_data()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$new(n, omega = 1, alpha = rep(1/n, n), rate = rep(1, n))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-omega"></a>}}
\if{latex}{\out{\hypertarget{method-omega}{}}}
\subsection{Method \code{omega()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$omega()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-alpha"></a>}}
\if{latex}{\out{\hypertarget{method-alpha}{}}}
\subsection{Method \code{alpha()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$alpha()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rate"></a>}}
\if{latex}{\out{\hypertarget{method-rate}{}}}
\subsection{Method \code{rate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$rate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$print(digits = max(3, getOption("digits") - 3), ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-get_params}{}}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$get_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_params"></a>}}
\if{latex}{\out{\hypertarget{method-init_params}{}}}
\subsection{Method \code{init_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$init_params(
  data,
  shape.init = c(1, 4, 16, 64, 256, 1024),
  scale.init = c(0.5, 1, 2),
  maxiter.init = 5,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params"></a>}}
\if{latex}{\out{\hypertarget{method-set_params}{}}}
\subsection{Method \code{set_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$set_params(params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-em"></a>}}
\if{latex}{\out{\hypertarget{method-em}{}}}
\subsection{Method \code{em()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$em(params, data, eps = 1e-08, ufactor = 1.01)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-llf"></a>}}
\if{latex}{\out{\hypertarget{method-llf}{}}}
\subsection{Method \code{llf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$llf(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CPHSRM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
